requirements.txt
python-dotenv==1.0.0
PyAudio==0.2.14
boto3==1.34.11
botocore==1.34.11
rx==3.2.0
aws_sdk_bedrock_runtime==0.0.2
smithy_aws_core==0.0.2
PyMuPDF==1.23.8
faiss-cpu==1.11.0
numpy==2.3.1
tqdm==4.66.2 